+++ src/full/Agda/TypeChecking/Serialise/Instances/Errors.hs
@@ -2,6 +2,10 @@
 {-# LANGUAGE CPP                      #-}
 {-# OPTIONS_GHC -fno-warn-orphans     #-}
 
+#if __GLASGOW_HASKELL__ <= 708
+{-# OPTIONS_GHC -fcontext-stack=30 #-}
+#endif
+
 module Agda.TypeChecking.Serialise.Instances.Errors where
 
 #if __GLASGOW_HASKELL__ <= 708
@@ -14,6 +18,7 @@ import Agda.TypeChecking.Serialise.Instances.Internal ()
 import Agda.TypeChecking.Serialise.Instances.Abstract ()
 
 import Agda.Syntax.Common
+import Agda.Syntax.Abstract.Name (ModuleName)
 import Agda.TypeChecking.Monad.Base
 import Agda.Interaction.Options
 import Agda.Termination.CutOff
@@ -28,25 +33,25 @@ import Agda.Utils.Lens
 import Agda.Utils.Impossible
 
 instance EmbPrj TCWarning where
-  icod_ (TCWarning a b c) = icode3' a b c
+  icod_ (TCWarning a b c) = icodeN' TCWarning a b c
 
-  value = value3 TCWarning
+  value = valueN TCWarning
 
 -- We don't need to serialise warnings that turn into errors
 instance EmbPrj Warning where
   icod_ (TerminationIssue a)         = __IMPOSSIBLE__
-  icod_ (UnreachableClauses a b)     = icode2 0  a b
+  icod_ (UnreachableClauses a b)     = icodeN 0 UnreachableClauses a b
   icod_ (CoverageIssue a b)          = __IMPOSSIBLE__
   icod_ (CoverageNoExactSplit a b)   = __IMPOSSIBLE__
   icod_ (NotStrictlyPositive a b)    = __IMPOSSIBLE__
   icod_ (UnsolvedMetaVariables a)    = __IMPOSSIBLE__
   icod_ (UnsolvedInteractionMetas a) = __IMPOSSIBLE__
   icod_ (UnsolvedConstraints a)      = __IMPOSSIBLE__
-  icod_ (OldBuiltin a b)             = icode1 1  a
-  icod_ EmptyRewritePragma           = icode0 2
-  icod_ UselessPublic                = icode0 3
-  icod_ (UselessInline a)            = icode1 4 a
-  icod_ (GenericWarning a)           = icode1 5 a
+  icod_ (OldBuiltin a b)             = icodeN 1 OldBuiltin a b
+  icod_ EmptyRewritePragma           = icodeN 2 EmptyRewritePragma
+  icod_ UselessPublic                = icodeN 3 UselessPublic
+  icod_ (UselessInline a)            = icodeN 4 UselessInline a
+  icod_ (GenericWarning a)           = icodeN 5 GenericWarning a
   icod_ (GenericNonFatalError a)     = __IMPOSSIBLE__
   icod_ (SafeFlagPostulate a)        = __IMPOSSIBLE__
   icod_ (SafeFlagPragma a)           = __IMPOSSIBLE__
@@ -56,7 +61,7 @@ instance EmbPrj Warning where
   icod_ SafeFlagNoPositivityCheck    = __IMPOSSIBLE__
   icod_ SafeFlagPolarity             = __IMPOSSIBLE__
   icod_ (ParseWarning a)             = __IMPOSSIBLE__
-  icod_ (DeprecationWarning a b c)   = icode3 6 a b c
+  icod_ (DeprecationWarning a b c)   = icodeN 6 DeprecationWarning a b c
 
   value = vcase valu where
       valu [0, a, b]     = valuN UnreachableClauses a b
@@ -69,14 +74,14 @@ instance EmbPrj Warning where
       valu _             = malformed
 
 instance EmbPrj Doc where
-  icod_ d = icode1' (render d)
+  icod_ d = icodeN' (undefined :: String -> Doc) (render d)
 
-  value = value1 text
+  value = valueN text
 
 instance EmbPrj a => EmbPrj (Closure a) where
-  icod_ (Closure a b c d e) = icode5' a b c d e
+  icod_ (Closure a b c d e) = icodeN' Closure a b c d e
 
-  value = value5 Closure
+  value = valueN Closure
 
 
 -- We are only serialising the parts of the environment and the state
@@ -84,17 +89,29 @@ instance EmbPrj a => EmbPrj (Closure a) where
 -- NOTE: If needed for other things, this might need to be changed!
 
 instance EmbPrj TCEnv where
-  icod_ TCEnv{..} = icode2' envCurrentModule (SerialisedRange envRange)
+  icod_ TCEnv{..} = icodeN' (undefined :: ModuleName -> SerialisedRange -> TCEnv)
+                            envCurrentModule (SerialisedRange envRange)
 
-  value = value2 $ \ a c -> initEnv
+  value = valueN $ \ a c -> initEnv
                  { envCurrentModule = a
                  , envRange = underlyingRange c
                  , envCall = Nothing
                  }
 
 instance EmbPrj TCState where
-  icod_ st = icode8' (st ^. stImports)
-                     (st ^. stModuleToSource)
+  icod_ st = icodeN' (undefined :: Signature -> PragmaOptions
+                                -> BuiltinThings PrimFun -> MetaStore
+                                -> InteractionPoints -> Signature
+                                -> BuiltinThings PrimFun -> TCState)
+                     (st ^. stImports)
+
+-- FNF 2017-06-13: Experiments when implementing serialising warnings
+-- suggested that we needed to serialise the ModuleToSource stored in
+-- the state for printing warnings stored in interface files, but this
+-- does not seem needed, and caused Issue 2592; hence the following
+-- line is now commented out.
+
+--                     (st ^. stModuleToSource)
                      (st ^. stPragmaOptions)
                      (st ^. stImportedBuiltins)
                      (st ^. stMetaStore)
@@ -102,9 +119,9 @@ instance EmbPrj TCState where
                      (st ^. stSignature)
                      (st ^. stLocalBuiltins)
 
-  value = value8 (\ a b c d e f g h ->
+  value = valueN (\ a c d e f g h ->
                      set stImports           a $
-                     set stModuleToSource    b $
+--                     set stModuleToSource    b $
                      set stPragmaOptions     c $
                      set stImportedBuiltins  d $
                      set stMetaStore         e $
@@ -114,40 +131,53 @@ instance EmbPrj TCState where
                      initState)
 
 instance EmbPrj InteractionId where
-  icod_ (InteractionId a) = icode1' a
+  icod_ (InteractionId a) = icodeN' InteractionId a
 
-  value = value1 InteractionId
+  value = valueN InteractionId
 
 instance EmbPrj PragmaOptions where
   -- TODO: only keep the options needed for displaying the warnings
-  icod_ (PragmaOptions a b c d e f g h i j k l m n o p q r s t u v w x y z) = icode26' a b c d e f g h i j k l m n o p q r s t u v w x y z
+  icod_ (PragmaOptions a b c d e f g h i j k l m n o p q r s t u v w x y z aa) =
+    icodeN' PragmaOptions a b c d e f g h i j k l m n o p q r s t u v w x y z aa
 
-  value = value26 PragmaOptions
+  value = valueN PragmaOptions
+
+instance EmbPrj WarningMode where
+  icod_ LeaveAlone        = icodeN 0 LeaveAlone
+  icod_ TurnIntoErrors    = icodeN 1 TurnIntoErrors
+  icod_ IgnoreAllWarnings = icodeN 2 IgnoreAllWarnings
+
+  value = vcase valu where
+    valu [0] = valuN LeaveAlone
+    valu [1] = valuN TurnIntoErrors
+    valu [2] = valuN IgnoreAllWarnings
+    valu _   = malformed
+
 
 instance EmbPrj PrimFun where
   -- don't need implementation for the warnings
-  icod_ (PrimFun a b _)= icode2' a b
+  icod_ (PrimFun a b r)= icodeN' (\ a b -> PrimFun a b r) a b
 
-  value = value2 (\ a b -> PrimFun a b __IMPOSSIBLE__)
+  value = valueN (\ a b -> PrimFun a b __IMPOSSIBLE__)
 
 instance EmbPrj CutOff where
-  icod_ (CutOff a) = icode1 0 a
-  icod_ DontCutOff = icode0 1
+  icod_ (CutOff a) = icodeN 0 CutOff a
+  icod_ DontCutOff = icodeN 1 DontCutOff
 
   value = vcase valu where
     valu [0, a] = valuN CutOff a
     valu [1]    = valuN DontCutOff
-    valu _         = malformed
+    valu _      = malformed
 
 instance EmbPrj MetaVariable where
-  icod_ (MetaVar a b c d _ _ _) = icode4' a b c d
+  icod_ (MetaVar a b c d r s t) = icodeN' (\ a b c d -> MetaVar a b c d r s t) a b c d
 
-  value = value4 $ \ a b c d ->
+  value = valueN $ \ a b c d ->
           MetaVar a b c d __IMPOSSIBLE__ __IMPOSSIBLE__ __IMPOSSIBLE__
 
 instance EmbPrj a => EmbPrj (Judgement a) where
-  icod_ (HasType a b) = icode2 0 a b
-  icod_ (IsSort a b)  = icode2 1 a b
+  icod_ (HasType a b) = icodeN 0 HasType a b
+  icod_ (IsSort a b)  = icodeN 1 IsSort a b
 
   value = vcase valu where
     valu [0, a, b] = valuN HasType a b
@@ -155,18 +185,18 @@ instance EmbPrj a => EmbPrj (Judgement a) where
     valu _         = malformed
 
 instance EmbPrj MetaPriority where
-  icod_ (MetaPriority a) = icode1' a
+  icod_ (MetaPriority a) = icodeN' MetaPriority a
 
-  value = value1 MetaPriority
+  value = valueN MetaPriority
 
 instance EmbPrj MetaInfo where
-  icod_ (MetaInfo a b c) = icode3' a b c
+  icod_ (MetaInfo a b c) = icodeN' MetaInfo a b c
 
-  value = value3 MetaInfo
+  value = valueN MetaInfo
 
 instance EmbPrj RunMetaOccursCheck where
-  icod_ RunMetaOccursCheck = icode0 0
-  icod_ DontRunMetaOccursCheck = icode0 1
+  icod_ RunMetaOccursCheck     = icodeN 0 RunMetaOccursCheck
+  icod_ DontRunMetaOccursCheck = icodeN 1 DontRunMetaOccursCheck
 
   value = vcase valu where
     valu [0] = valuN RunMetaOccursCheck
@@ -174,12 +204,12 @@ instance EmbPrj RunMetaOccursCheck where
     valu _   = malformed
 
 instance EmbPrj InteractionPoint where
-  icod_ (InteractionPoint a b c _) = icode3' a b c
+  icod_ (InteractionPoint a b c r) = icodeN' (\ a b c -> InteractionPoint a b c r) a b c
 
-  value = value3 (\ a b c -> InteractionPoint a b c __IMPOSSIBLE__)
+  value = valueN (\ a b c -> InteractionPoint a b c __IMPOSSIBLE__)
 
 instance EmbPrj ModuleParameters where
-  icod_ (ModuleParams a) = icode1' a
+  icod_ (ModuleParams a) = icodeN' ModuleParams a
 
-  value = value1 ModuleParams
+  value = valueN ModuleParams